@@ -8,22 +8,24 @@
     "vite:build": "vite build",
     "tailwind:dev": "npx tailwindcss -i ./src/styles/index.css -o ./src/styles/tailwindOutput.css --watch",
     "tailwind:build": "npx tailwindcss -o ./src/styles/tailwindOutput.css --minify",
-    "dev": "start npm run tailwind:dev && vite",
+    "dev": "vite",
     "build": "npm run tailwind:build && npm run vite:build",
     "preview": "vite preview"
   },
   "dependencies": {
     "@vitejs/plugin-react": "^1.0.7",
-    "firebase": "^9.10.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.4.0",
-    "react-loading-screen": "0.0.17",
-    "react-router-dom": "^6.2.2",
-    "tailwindcss": "^3.0.23"
+    "firebase": "^10.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.11.0",
+    "react-loading-screen": "^0.0.17",
+    "react-router-dom": "^6.15.0",
+    "tailwindcss": "^3.3.3"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^1.0.7",
+    "autoprefixer": "^10.4.15",
+    "postcss": "^8.4.29",
     "vite": "^2.8.0"
   }
 }
