@@ -1,6 +1,6 @@
-import { initializeApp } from "firebase/app"
-import { getFirestore } from "@firebase/firestore";
-import { getAuth } from "@firebase/auth";
+import { initializeApp } from "@firebase/app"
+import { getFirestore, initializeFirestore } from "@firebase/firestore";
+import { getAuth, initializeAuth } from "@firebase/auth";
 console.log(import.meta.env)
 const firebaseConfig = {
   apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
@@ -12,5 +12,5 @@ const firebaseConfig = {
 }
 
 export const app = initializeApp(firebaseConfig);
-export const db = getFirestore(app);  
 export const auth = getAuth(app);
+export const db = getFirestore(app);  
